extract person selector component
